feat: create sign up page in app dir
